[refactor] Account 엔티티 balance 컬럼 타입 변경 - Long -> BigDecimal - 오버플로우 방지 및 소수점 연산 대비
[chore] H2Dialect 설정 추가 및 그 외 설정 내용 설명 추가

+++ src/main/java/com/test/learningtx/repository/AccountRepository.java
@@ -8,6 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -29,5 +30,5 @@ public interface AccountRepository extends JpaRepository<Account, Long> {
 
     Optional<Account> findByName(String name);
 
-    List<Account> findByBalanceBetween(Long minBalance, Long maxBalance);
+    List<Account> findByBalanceBetween(BigDecimal minBalance, BigDecimal maxBalance);
 }
